+++ R/calculus.R
@@ -183,6 +183,8 @@ prodf <- function(f, var, lower, upper, doit = TRUE) {
 #'   intf(1/x, x, 1, 10, doit = FALSE)
 #'   intf(1/x, x)
 #'   intf(1/x, x, doit = FALSE)
+#'   intf(exp(-x^2/2), x, -Inf, Inf)
+#'   intf(exp(-x^2/2), x, -Inf, Inf, doit = FALSE)
 #' }
 #' 
 #' @concept calculus